backend/requirements.txt
flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9